new: test full pipeline, move id/timestamp assignment into parser.
